fixed ton of things, almost complete whole reading part

+++ JavaDeobfuscator/JavaAsm/IO/ConstantPoolEntries/LongEntry.cs
@@ -14,7 +14,7 @@ namespace JavaDeobfuscator.JavaAsm.IO.ConstantPoolEntries
 
         public LongEntry(Stream stream)
         {
-            Value = Binary.BigEndian.ReadInt32(stream);
+            Value = Binary.BigEndian.ReadInt64(stream);
         }
 
         public override EntryTag Tag => EntryTag.Long;
